# cmd/swag/docs/swagger.yaml
basePath: /v1/2025
definitions:
...
  model.ChangePasswordInput:
    properties:
      confirmpassword:
        type: string
      newpassword:
        type: string
      oldpassword:
        type: string
    type: object
  model.License:
    properties:
      dateend:
        example: "2025-03-30 11:22:30"
        type: string
      datestart:
        example: "2025-03-28 11:22:30"
        type: string
      role_id:
        example: 2b796313-1134-44b3-b527-2c27d41a1624
        type: string
    required:
    - dateend
    - datestart
    - role_id
    type: object
  model.LoginInput:
    properties:
      password:
        example: "123"
        type: string
      username:
        example: phananhtu1998
        type: string
    required:
    - password
    - username
    type: object
  model.MenuInput:
    properties:
      id:
        type: string
      menu_group_name:
        example: /api/dashboard
        type: string
      menu_icon:
        example: database
        type: string
      menu_level:
        example: 1
        type: integer
      menu_name:
        example: Dashboard
        type: string
      menu_number_order:
        example: 1
        type: integer
      menu_parent_id:
        type: string
      menu_url:
        example: /dashboard
        type: string
    required:
    - menu_group_name
    - menu_icon
    - menu_level
    - menu_name
    - menu_number_order
    - menu_url
    type: object
  model.Role:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      is_licensed:
        type: boolean
      role_left_value:
        type: integer
      role_max_number:
        type: integer
      role_name:
        type: string
      role_right_value:
        type: integer
      updated_by:
        type: string
    type: object
...
  response.ErrorResponseData:
    properties:
      code:
        description: status code
        type: integer
      detail:
        description: dữ liệu được return
      error:
        description: Thông báo lỗi
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: status code
        type: integer
      data:
        description: dữ liệu được return
      message:
        description: Thông báo lỗi
        type: string
    type: object
host: localhost:8002
info:
  contact:
    email: teamaithacoindustries@gmail.com
    name: TEAM AI
    url: github.com/phananhtu1998/go-backend-api
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: github.com/phananhtu1998/go-backend-api
  title: API DOCUMENTATION PARKING BACKEND
  version: 1.0.0
paths:
...
  /auth/change_password:
    post:
      consumes:
      - application/json
      description: ChangePassword
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      summary: ChangePassword
      tags:
      - Authenticate
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Login
      tags:
      - Authenticate
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Authenticate
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: RefreshToken
      parameters:
      - description: Refresh Token
        in: header
        name: RefreshToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - ApiKeyAuth: []
      summary: RefreshToken
      tags:
      - Authenticate
  /license/create_license:
    post:
      consumes:
      - application/json
      description: API này cho phép mới license
      parameters:
      - description: Thông tin License cần tạo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.License'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Tạo mới license
      tags:
      - License
  /menu/create_menu:
    post:
      consumes:
      - application/json
      description: API tạo menu trong hệ thống
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.MenuInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Tạo menu
      tags:
      - Menu
  /menu/delete/{id}:
    delete:
      consumes:
      - application/json
      description: API này xóa một menu dựa trên ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Xóa menu
      tags:
      - Menu
  /menu/get_all_menu:
    get:
      consumes:
      - application/json
      description: API này trả về danh sách tất cả menu trong hệ thống
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Lấy danh sách tất cả menu
      tags:
      - Menu
  /menu/get_menu_by_id/{id}:
    get:
      consumes:
      - application/json
      description: API này trả về menu theo ID
      parameters:
      - description: ID menu cần lấy
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Lấy menu theo ID
      tags:
      - Menu
  /menu/update_multiple_menu:
    put:
      consumes:
      - application/json
      description: API này cập nhật danh sách menu dựa trên danh sách ID
      parameters:
      - description: Danh sách menu cần cập nhật
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/model.MenuInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Cập nhật nhiều menu
      tags:
      - Menu
  /role/create_role:
    post:
      consumes:
      - application/json
      description: Api tạo vai trò trong hệ thống
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      summary: Tạo vai trò
      tags:
      - Role
  /role/delete_role/{id}:
    delete:
      consumes:
      - application/json
      description: API xóa mềm role trong hệ thống
      parameters:
      - description: ID của role cần xóa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Xóa role
      tags:
      - Role
  /role/get_all_roles:
    get:
      consumes:
      - application/json
      description: Api lấy danh sách role trong hệ thống
      parameters:
      - description: 'Số trang (mặc định: 1)'
        in: query
        name: page
        type: integer
      - description: 'Số lượng mỗi trang (mặc định: 20)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      summary: Lấy danh sách role
      tags:
      - Role
  /role/get_role_by_id/{id}:
    get:
      consumes:
      - application/json
      description: API này trả về role theo ID
      parameters:
      - description: ID role cần lấy
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      summary: Lấy role theo ID
      tags:
      - Role
...
security:
- "":
  - BearerAuth
  - ApiKeyAuth
securityDefinitions:
  ApiKeyAuth:
    description: Nhập API Key vào header để xác thực API
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Nhập Bearer Token vào header. Ví dụ: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"